refine existing ui ( spacing, colors, typography)
Fixed compile warnings.
Returned the search's opacity to 100%.
Centered the guesses element.
Converted key-up event to key-down.
As i nstructed i tried to create a system for the colors, spacings and font-sizes which made the overall code more organised. It can be subject to imporvement but for now I think is counter-intuitive to persist.