Help-modal improvements
Replaced all the widths and heights to pixels.
Made it so clicking on the info button opens the help modal.
Changed the close button to an actual icon.
Altered the page spacing so that the components have more breathing room.
Cached the seen Modal value so that the user sees it only the first time they visit the page or if they click the info button.
Added a simple animation to the modal appearing.