Authentication
created sign-in folder with an index.vue
on the pages folder and implemented functionality to navigate to using the sign in button on the Header.
Created sign-in card component which has the sign in with google or apple button.
Put the show/hide help modal functionality on a UseModal composable