Supabase implementation
Installed supabase.
Created useAuth composable to handle signing-in, signing-out and to store the user state.
Created "User.vue" component to display the user information if one is singed-in.
Deleted "sign in with apple" button since getting an apple developer account costs 100$/year and we decided to drop it at least for the time being.
Minor css changes.