comply to review changes
Deleted unused ref from "useModal" composable.
Added an svg next to the logout button.
Changed the username displayed on the user info as to only display the first name and not the full name.
Changed the props the Button component takes as to be read easier and comply with the css standard practises.
Changed the Header and SigninCard components to give props to their respective buttons according to the changes made.
Erased unnecessary dependecies installed by mistake.