fixed navigation issues.
Modified the signInWithGoogle function to redirect to the page's domain or to localhost depending on being on development or production.
The app understands if its on develoment from a .env file which a value "isDevelopment=true" is stored.
Erased the "League Grid" button and entered the functionality to go to '/' on click on the upper left logo of the header component.
Entered a check for user existence so that if a user is logged in, they cannot go to the 'sign-in' page, they are instead redirected to '/'. In addition I made the page not render the signInCard component in case there already is a User signed-in