Help modal
Moved the help icon to the HelpModal component along with the show/hide modal functionality.
Made the help so that is always visible on the header and always clickable.
Created an empty section with id"help-modal" on the app.vue file. This was done because in between different pages the help modal was dysfunctional and could not otherwise exist in both/all pages.
Made it so that if you have selected a cell and click another one, the second will be immediatly selected, as opposed to befoe where we stopped the event propagation.
If another component besides a cell is clicked, the event propagation will stop.