Code layout improvement
Moved the focusList function to th navigateList.ts file since this is where it is used and we dont have to keep 2 separate files storing two small functions each.
Move Game.vue component's functionality to UseGame composable.
Moved SearchResults .vuecomponent's functionality to UseNavigation composable.
Created media queries to change the css in the images and metadata of the cells.
Change the Button component's padding.
Added an extra boolean prop taken by the Cell component to make sure that when the guesses are down to zero
the hovered cells are not highlighted