Minor changes.
Made score be displayed with one decimal point and only if the rarityScore is a real number.
Turned the bucket url into a public env variable called with useNuxtConfig.
Passed the guesses variable to the Grid component so that it can check if there are any gueesses left before selecting a cell.
Changed the condition on handle champion chosen to be score>=0 since the function may return 0 if the answer is correct but the rarity score is zero.