Bug fixes
1. Cell.vue
corrected typo  on hoverable class.
2. Grid.
 change the condition on onCellClick so it wont execute if the cell is already populated.
 3. Search.vue.
 Put a v-if on search component so that it wont appear if the user deletes the search value.
 4. Changed the type of the prop listItems to be an array of {name,id} values
and changed the condition of the onmounted function so it wont execute the handleChampionChosen funtion if there is no focusedChoice value.
5. navigateList.ts
change the results type on the getFocusedChoice function and made ut return it's own value as a catch all return so that if the conditions arrent fullfilled, the foccused choise value remains the same.