pinia installation/config
installed the pinia package and changed the configuration on nuxt.config  as to auto import the pinia stores from the stores folder.
Created usePuzzleStore to handle fetching the puzzle data.
Moved the logic to fetch data from supabase to the server, created a server middleware to extract the auth token and the user from the cookies