NavBar
Created a NavBar component to hande the routing between the user's puzzle, the page showing most-popular answers for each cell and the leaderBoard.
Created a most-popular page.
Changed the game.vue component to take its data from the puzzle store instead of the useGame composable and to check if there's a logged in user so it will fetch the local game on mounted.
Fixed warning on sign-in page that occured because the buttonClass prop of the button had not been set to optional.