Code improvements
Deleted the most-popular game grid and cells, modified the already existing components to be used on the most-popular page.
Changed the error handling so that the srver won't throw fatal errors and crash the server.
Deleted the PuzzleMetadata  values and moved the champion ids and rarity scores to the champion body and the possible aanswers to the champion info.